Arreglos varios. Preparación para V2 de página
