+++ src/modules/Home/components/ApplicationsSection/ApplicationsSection.css
@@ -1,17 +1,15 @@
 .applications-section{
     width: 100%;
     display: grid;
-    grid-template-columns: repeat(3, 1fr);
+    grid-template-columns: repeat(2, 1fr);
     gap: 20px;
 }
 
 .applications-section-item{
     width: 100%;
-    aspect-ratio: 3 / 4;
     border-radius: 10px;
-    display: grid;
-    grid-template-rows: repeat(5, 1fr);
-    /* border: 1px solid #d0d5dd; */
+    display: flex;
+    flex-direction: column;
     box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
 }
 
@@ -34,8 +32,8 @@
 
 .applications-section-item-img{
     padding: 10px;
-    width: 100%;
-    height: 100%;
+    width: 390px;
+    height: 280px;
     grid-row: 1 / 5;
     border-radius: 10px;
     border-radius: 10px;
@@ -53,6 +51,23 @@
     transition: 0.5s;
     transform: scale(1.1);
     border-radius: 10px;
+}
+
+.applications-section-item-feature{
+    transition: 0.5s;
+    width: fit-content;
+    align-self: flex-end;
+    margin: 20px;
+    padding: 5px 10px;
+    border-radius: 10px;
+    color: var(--background_color);
+    background-color: rgba(var(--primary_color), 0.8);
+}
+
+.applications-section-item-feature:hover{
+    cursor: default;
+    transition: 0.5s;
+    background-color: rgba(var(--primary_color), 1);
 }
 
 @media (max-width: 440px){
