+++ src/modules/Home/components/ApplicationsSection/ApplicationsSection.jsx
@@ -1,46 +1,77 @@
 import "./ApplicationsSection.css";
 
-import galleryImg1 from "../../../../assets/gallery_img_1.jpg";
-import galleryImg2 from "../../../../assets/gallery_img_2.jpg";
-import galleryImg3 from "../../../../assets/gallery_img_3.jpg";
+import applicationsSectionImg1 from "../../../../assets/homePage/ApplicationsSection/1.jpg";
+import applicationsSectionImg2 from "../../../../assets/homePage/ApplicationsSection/2.jpg";
+import applicationsSectionImg3 from "../../../../assets/homePage/ApplicationsSection/3.jpg";
+import applicationsSectionImg4 from "../../../../assets/homePage/ApplicationsSection/4.jpg";
+import applicationsSectionImg5 from "../../../../assets/homePage/ApplicationsSection/5.jpg";
+import applicationsSectionImg6 from "../../../../assets/homePage/ApplicationsSection/6.jpg";
+import applicationsSectionImg7 from "../../../../assets/homePage/ApplicationsSection/7.jpg";
+import applicationsSectionImg8 from "../../../../assets/homePage/ApplicationsSection/8.jpg";
 
 export default function ApplicationsSection() {
   return (
     <div className="applications-section">
       <div className="applications-section-item">
         <div className="applications-section-item-img">
-          <img src={galleryImg1} />
+          <img src={applicationsSectionImg4} />
         </div>
-        
         <div className="applications-section-item-text">
-          <span className="applications-section-item-title">Electronica</span>
+          <span className="applications-section-item-title">Componentes para el Sector Eléctrico</span>
           <span className="applications-section-item-description">
-            Electronica
+            Fabricación de tapas, cubiertas y palancas de fuerza para tableros eléctricos industriales.
           </span>
         </div>
       </div>
       <div className="applications-section-item">
+        <div className="applications-section-item-img">
+          <img src={applicationsSectionImg7} />
+        </div>
         <div className="applications-section-item-text">
-          <span className="applications-section-item-title">Fixtures</span>
+          <span className="applications-section-item-title">Fixtures y Herramientas para Ensambles</span>
           <span className="applications-section-item-description">
-            Fixtures
+            Diseño e impresión de fixtures personalizados para agilizar procesos de ensamblaje en líneas de producción.
           </span>
         </div>
+        <span className="applications-section-item-feature">
+          Nuevo
+        </span>
+      </div>
+      <div className="applications-section-item">
         <div className="applications-section-item-img">
-          <img src={galleryImg2} />
+          <img src={applicationsSectionImg8} />
+        </div>
+        <div className="applications-section-item-text">
+          <span className="applications-section-item-title">Piezas de Reemplazo</span>
+          <span className="applications-section-item-description">
+            Producción de repuestos difíciles de conseguir o descontinuados mediante ingeniería inversa.
+          </span>
         </div>
       </div>
       <div className="applications-section-item">
-        
-      <div className="applications-section-item-text">
-          <span className="applications-section-item-title">Carcasas</span>
+        <div className="applications-section-item-img">
+          <img src={applicationsSectionImg5} />
+        </div>
+        <div className="applications-section-item-text">
+          <span className="applications-section-item-title">Producción en Pequeñas Series</span>
           <span className="applications-section-item-description">
-            Carcasas
+            Fabricación de lotes pequeños sin necesidad de moldes costosos, ideal para pruebas y validaciones antes de producción en masa.
           </span>
         </div>
+      </div>
+      <div className="applications-section-item">
         <div className="applications-section-item-img">
-          <img src={galleryImg3} />
+          <img src={applicationsSectionImg1} />
         </div>
+        <div className="applications-section-item-text">
+          <span className="applications-section-item-title">Prototipos y Desarrollo de Productos</span>
+          <span className="applications-section-item-description">
+            Creación de prototipos funcionales y modelos de prueba antes de la producción final.
+          </span>
+        </div>
+        <span className="applications-section-item-feature">
+          Nuevo
+        </span>
       </div>
     </div>
   );
