+++ src/modules/Materials/components/MaterialDisplay/IndividualMaterial/IndividualMaterial.css
@@ -1,0 +1,248 @@
+.individual_material{
+  margin-top: var(--nav_bar_height);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+}
+
+.individual_material_title{
+    font-size: 100px;
+    line-height: 100px;
+    font-weight: 600;
+    background: linear-gradient(
+        to right,
+        #5dc6ff 20%,
+        #00a3fa 30%,
+        #00a3fa 70%,
+        #5dc6ff 80%
+    );
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+    text-fill-color: transparent;
+    background-size: 500% auto;
+    animation: textShine 5s ease-in-out infinite alternate;
+}
+
+.individual_material_item{
+    display: flex;
+    gap: 40px;
+}
+
+.individual_material_item_material{
+    width: 400px;
+    height: 400px;
+    border-radius: 50%;
+    /* animation: 3s skeleton infinite ease-in-out; */
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.individual_material_item_material img{
+    width: 100%;
+}
+
+.individual_material_item_left{
+    width: 33%;
+    display: flex;
+    flex-direction: column;
+    align-items: end;
+    gap: 20px;
+    text-align: end;
+    justify-content: center;
+}
+
+.individual_material_item_right{
+    width: 33%;
+    display: flex;
+    flex-direction: column;
+    align-items: start;
+    gap: 20px;
+    text-align: start;
+    justify-content: center;
+}
+
+.individual_material_icon{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    gap: 20px;
+}
+
+.individual_material_icon_container{
+    position: relative;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.individual_material_icon_bg{
+    width: 70px;
+}
+
+.individual_material_icon_front{
+    position: absolute;
+    width: 25px;
+    filter: invert(1);
+    animation: 4s icon infinite;
+}
+
+.individual_material_spec{
+    width: max-content;
+    max-width: 600px;
+    font-size: 25px;
+    color: rgba(var(--text_color), 0.9);
+}
+
+.individual_material_spec_highlight{
+    color: rgba(var(--primary_color), 1);
+}
+
+.tooltip_container {
+    position: relative;
+    display: inline-block;
+}
+
+.individual_material_icon_container .tooltiptext_container {
+    visibility: hidden;
+    min-width: 180px;
+    background-color: white;
+    border: 3px solid black;
+    text-align: center;
+    padding: 5px 0;
+    border-radius: 6px;
+    margin-left: 7px;
+    margin-bottom: 150px;
+    position: absolute;
+    z-index: 1;
+    padding: 10px;
+    font-size: 20px;
+}
+
+.tooltiptext_container_tail{
+    width: 15px;
+    height: 15px;
+    background-color: rgba(var(--primary_color), 1);
+    border: 3px solid black;
+    position: absolute;
+    left: 43%;
+    bottom: -10px;
+    transform: rotate(45deg);
+    z-index: 1;
+}
+
+.individual_material_icon_container:hover .tooltiptext_container {
+    visibility: visible;
+}
+
+.individual_material_description{
+    width: 1400px;
+    margin-top: 50px;
+    font-size: 20px;
+    text-align: center;
+    color: rgba(var(--text_color), 0.7);
+}
+
+.individual_material_button{
+    margin-top: 40px;
+    width: 100%;
+    background-color: rgba(var(--primary_color), 1);
+    color: #fff;
+    text-decoration: none;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    text-align: center;
+    border-radius: 10px;
+    font-size: 25px;
+    padding: 8px 0;
+    animation: 2s button infinite ease-in-out;
+}
+
+@keyframes
+skeleton {
+    0% {background-color:rgb(223, 223, 223); left:0px; top:0px;}
+    50% {background-color:rgb(173, 173, 173); left:0px; top:0px;}
+    100% {background-color:rgb(223, 223, 223); left:0px; top:0px;}
+}
+  
+@keyframes
+icon {
+    0% {transform: scale(1);}
+    50% {transform: scale(1.4);}
+    100% {transform: scale(1);}
+}
+
+@keyframes
+textShine {
+    0% {
+        background-position: 0% 50%;
+    }
+    100% {
+        background-position: 100% 50%;
+    }
+}
+
+@keyframes
+button {
+    0% {transform: scale(1);}
+    50% {transform: scale(1.05);}
+    100% {transform: scale(1);}
+}
+
+@media (max-width: 440px) {
+    .individual_material{
+        align-items: center;
+        justify-content: center;
+        gap: 30px;
+    }
+    .individual_material_title{
+        font-size: 80px;
+        margin-bottom: -15px;
+    }
+    .individual_material_item{
+        flex-direction: column;
+        width: auto;
+        justify-content: center;
+        align-items: center;
+        gap: 20px;
+    }
+    .individual_material_item_left{
+        width: calc(100% - 20px);
+        padding: 0 10px;
+    }
+    .individual_material_item_right{
+        width: calc(100% - 20px);
+        padding: 0 10px;
+    }
+    .individual_material_icon{
+        gap: 10px;
+    }
+    .individual_material_icon_bg{
+        width: 50px;
+    }
+    .individual_material_icon_front{
+        width: 20px;
+    }
+    .individual_material_spec{
+        font-size: 18px;
+    }
+    .individual_material_icon_container .tooltiptext_container{
+        visibility: hidden;
+        display: none;
+    }
+    .individual_material_description{
+        width: auto;
+        text-align: left;
+        margin-top: 0px;
+    }
+    .individual_material_button{
+        margin-top: 20px;
+        width: auto;
+        font-size: 32px;
+        padding: 10px 20px;
+        width: calc(100% - 80px);
+    }
+}
